+++ README.md
@@ -10,8 +10,8 @@
 
 ### 요구사항
 
-- [ ] 본인만의 기준으로 일관된 코드를 작성해주세요.
-- [ ] 기능 단위로 나누어 커밋을 해주세요.
+- [x] 본인만의 기준으로 일관된 코드를 작성해주세요.
+- [x] 기능 단위로 나누어 커밋을 해주세요.
 - [x] 로그인 기능을 만들어주세요.
   - [x] ID는 아래의 조건에 맞게 동작해요.
     - [x] ID는 반드시 입력 되어야 해요.
@@ -25,5 +25,5 @@
            (ex. PW를 입력해주세요., PW는 최소 8글자 이상이어야 합니다.)
   - [x] 로그인 버튼은 아래 조건에 맞게 동작해요.
     - [x] ID의 조건과 PW 조건이 모두 충족 되었을 때에만 버튼이 활성화 돼요.
-- [ ] 만약 위 로직을 구현하면서 커스텀 훅을 사용하지 않았다면 재사용을 고려해 커스텀 훅으로 로그인 기능을 리팩터링하세요.  
+- [x] 만약 위 로직을 구현하면서 커스텀 훅을 사용하지 않았다면 재사용을 고려해 커스텀 훅으로 로그인 기능을 리팩터링하세요.  
        (**`LoginFormSection`** 에 직접적으로 `useState`를 사용하지 않는 방향으로 만들어보세요.)
