+++ src/hooks/useLoginForm.ts
@@ -1,0 +1,46 @@
+import { useState } from 'react';
+
+const useLoginForm = () => {
+  const [email, setEmail] = useState('');
+  const [emailError, setEmailError] = useState('');
+  const [password, setPassword] = useState('');
+  const [passwordError, setPasswordError] = useState('');
+
+  const validateEmail = (value: string) => {
+    if (!value) {
+      setEmailError('ID를 입력해주세요.');
+    } else if (!/^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/.test(value)) {
+      setEmailError('ID는 이메일 형식으로 입력해주세요.');
+    } else {
+      setEmailError('');
+    }
+  };
+
+  const validatePassword = (value: string) => {
+    if (!value) {
+      setPasswordError('PW를 입력해주세요.');
+    } else if (value.length < 8) {
+      setPasswordError('PW는 최소 8글자 이상이어야 합니다.');
+    } else {
+      setPasswordError('');
+    }
+  };
+
+  const isEmailValid = !emailError && email.length > 0;
+  const isPasswordValid = !passwordError && password.length >= 8;
+  const isButtonValid = isEmailValid && isPasswordValid;
+
+  return {
+    email,
+    setEmail,
+    emailError,
+    validateEmail,
+    password,
+    setPassword,
+    passwordError,
+    validatePassword,
+    isButtonValid,
+  };
+};
+
+export default useLoginForm;
